services/google-observer/requirements.txt
kafka-python
google-cloud-storage
google-cloud-pubsub
